requirements.txt
lastversion
PyGithub
tabulate
pickledb